backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-cors-headers==4.5.0
django-geoposition==0.3.0
django-parler==2.3
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
numpy==2.1.2
packaging==24.1
pillow==10.4.0
pip-review==1.3.0
postgis==1.0.4
psycopg==3.2.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pygrib==2.1.6
pyproj==3.7.0
python-dotenv==1.0.1
requests==2.32.3
scipy==1.14.1
sqlparse==0.5.1
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.3
